upgrade deneb-ui, fix a bug in home
